IBX-538: Renamed Installer bundle RepositoryInstaller
