+++ src/bundle/RepositoryInstaller/DependencyInjection/IbexaRepositoryInstallerExtension.php
@@ -4,7 +4,7 @@
  * @copyright Copyright (C) Ibexa AS. All rights reserved.
  * @license For full copyright and license information view LICENSE file distributed with this source code.
  */
-namespace Ibexa\Bundle\Installer\DependencyInjection;
+namespace Ibexa\Bundle\RepositoryInstaller\DependencyInjection;
 
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\Config\FileLocator;
