+++ src/bundle/RepositoryInstaller/Installer/Installer.php
@@ -4,7 +4,7 @@
  * @copyright Copyright (C) Ibexa AS. All rights reserved.
  * @license For full copyright and license information view LICENSE file distributed with this source code.
  */
-namespace Ibexa\Bundle\Installer\Installer;
+namespace Ibexa\Bundle\RepositoryInstaller\Installer;
 
 /**
  * Interface Installer.
